Split classes into seperate files under model subfolder
