+++ app/model/character_details.py
@@ -1,0 +1,8 @@
+from pydantic import BaseModel
+
+class Character_Details(BaseModel):
+    id: int
+    name: str
+    race: str
+    char_class: str
+    level : int
