+++ app/model/character_list.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+class Character_List(BaseModel):
+    id: int
+    belongs_to:str
