+++ app/model/character_skills.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+class Character_Skills(BaseModel):
+    id: int
+    skill: str
+    value: int
