+++ app/model/user.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+class User(BaseModel):
+    alias: str
+    password_hash: str
+    email: str
